+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   * Create `GenerateRulesError`
   * Create `Options` struct to pass some parameters to `Process`
   * Add `GenerateRules` option to `Options`
+  * Unexport `description.New` and `description` `.NextRecord` method
 
 ## 0.2 (15th July 2017)